requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
gunicorn==21.2.0
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.4
aiofiles==23.1.0
python-docx==0.8.11
azure-storage-blob==12.19.1
azure-search-documents==11.4.0
openai==1.30.1
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.3.0
PyMuPDF==1.24.9